packages/Avenue/Avenue-0.2.2.tar.gz/Avenue-0.2.2/src/avenue/avenue.py
# -*- coding: utf-8 -*-
from __future__ import absolute_import, division, print_function

from .processors import Match
from .processors import MatchError
from .processors import MethodProcessor
from .processors import PathProcessor
import logging
import operator

PROCESSORS = [
    PathProcessor('path'),
    MethodProcessor('method', optional=True),
]

logger = logging.getLogger('avenue')


class AvenueException(Exception):
    pass


class Skip(AvenueException):
    pass


class NotFound(AvenueException):
    pass


class Avenue(object):
    def __init__(self, processors=None):
        self.routes = list()
        self.processors = processors or PROCESSORS

    def attach(self, **kwargs):
        def inner(func):
            self.routes.append(Route(self, kwargs, func))
            return func

        return inner

    def match(self, **kwargs):
        values = [[p, kwargs.get(p.name)] for p in self.processors]
        matches = list()

        for route in self.routes:
            try:
                match = Match()
                for processor, value in values:
                    match.update(processor(value, route))
            except MatchError:
                pass
            else:
                matches.append([route, match])

        if len(matches):
            size = max(map(len, [m[1].score for m in matches]))
            matches = list(filter(lambda m: len(m[1].score) == size, matches))
            matches.sort(key=operator.itemgetter(1))

        logger.debug('Matched %s', [v[1] for v in values])
        for match in matches:
            logger.debug('%s', match)

        return matches

    def solve(self, **kwargs):
        wrap = kwargs.pop('wrap', None)

        for path, match in self.match(**kwargs):
            try:
                if callable(wrap):
                    callback = wrap(path.func)
                else:
                    callback = path.func

                return callback(**match.kwargs)
            except Skip:
                logger.debug('Callback raised Skip for match %s', match)

        raise NotFound()

    def blueprint(self, **kwargs):
        return Blueprint(self, **kwargs)

class Route(object):
    def __init__(self, avenue, kwargs, func):
        self.avenue = avenue
        self.kwargs = kwargs
        self.func = func


class Blueprint(object):
    def __init__(self, avenue, **kwargs):
        self.avenue = avenue
        self.kwargs = kwargs

    def attach(self, **kwargs):
        for processor in self.avenue.processors:
            name = processor.name
            value = kwargs.get(name)
            if value is not None:
                kwargs[name] = processor.blueprint(
                    value, self.kwargs.get(name))

        return self.avenue.attach(**kwargs)
